Crate problem025

Problem 25 (1000-digit Fibonacci number)

Solution detail

We don't actually need to calculate any Fibonacci numbers for this, which is a relief, since it's nice to avoid dealing with 1000-digit numbers if we can help it. Since Fn is the closest integer to (phi ^ n) / √5, we just need to find the first n for which:

n * log_10(ϕ) > 1000 + log_10(√5) - 1

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the index of the first Fibonacci number with at least n digits.